+++ packages/payloadset/packages/xns/plugins/record/src/helpers/name/spec/tsconfig.json
