+++ packages/payloadset/packages/xns/plugins/record/src/helpers/validation/validators/validators.ts
@@ -1,7 +1,7 @@
 import { isModuleName } from '@xyo-network/module-model'
 import type { PayloadValidationFunction, WithSources } from '@xyo-network/payload-model'
 
-import type { Domain } from '../../../Domain/index.ts'
+import type { Domain } from '../../../../../../../../../payload/packages/xns/plugins/record/src/Domain/index.ts'
 
 export const domainRegistrationCasingValidator: PayloadValidationFunction<WithSources<Domain>> = (
   domainRegistration: WithSources<Domain>,
